+++ class-components/src/App.css
@@ -1,11 +1,39 @@
+html {
+  height: 100%;
+}
+
+body {
+  height: 100%;
+  margin: 0;
+}
+
 #root {
   position: relative;
 
-  max-width: 1280px;
+  height: 100%;
   margin: 0 auto;
-  padding: 2rem;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
 
   text-align: center;
+}
+
+.container {
+  flex: 1 1 auto;
+  margin: 0 auto;
+  padding: 20px 0;
+  max-width: 1280px;
+}
+
+header,
+footer {
+  padding: 1rem 0;
+
+  color: white;
+  font-size: 20px;
+
+  background-color: rgb(82, 82, 82);
 }
 
 button {
@@ -20,8 +48,8 @@ button {
 
 button.button-error {
   position: absolute;
-  top: 0;
-  right: 0;
+  top: 5px;
+  right: 5px;
 
   border-color: black;
 
@@ -65,6 +93,8 @@ input {
 .list-item {
   padding: 10px 0;
   width: 450px;
+
+  cursor: pointer;
 }
 
 .list-item:not(:last-child) {
