+++ class-components/src/pages/index.ts
@@ -1,0 +1,4 @@
+import MainPage from './MainPage';
+import Page404 from './Page404';
+
+export { MainPage, Page404 };
